Prevent calling enif_make_atom with NULL env
It is not causing problems right now because enif_make_atom implementation
does not actually use this env. However, this behaviour is not documented
an we should not rely on it.

The env in unifex_get_pid_by_name may be NULL if called from created thread
(documentation states enif_whereis_pid should get NULL then,
but process-independent env also happens to work)